+++ frame/ethink/src/lib.rs
@@ -28,6 +28,8 @@
 #![allow(clippy::comparison_chain, clippy::large_enum_variant)]
 
 #[cfg(all(feature = "std", test))]
+mod mock;
+#[cfg(all(feature = "std", test))]
 mod tests;
 
 use frame_support::{
